use data table + update design tokens
